Finished pdf, more on searchgroup

+++ src/main/java/com/rihanna/neo4j/eg3/service/NutritionCalculatorService.java
@@ -28,7 +28,7 @@ public class NutritionCalculatorService {
 		return result;
 	}
 	
-	private float calculateTotal(float raw, IngredientQuantity ingQuan) {
-		return ((raw * ingQuan.getQuantity()) / ingQuan.getIngredient().getPerQuantity());
+	private Integer calculateTotal(Integer raw, IngredientQuantity ingQuan) {
+		return Float.valueOf(((raw * ingQuan.getQuantity()) / ingQuan.getIngredient().getPerQuantity())).intValue();
 	}
 }
